sunxi: Support UART1 on PG pins for the V3/S3
The V3/S3 has extra pins for UART1 on PG pins, which are not present
on the V3s. Add support for them and make them the default UART1 pins
on the V3. Note that PE pins are also avaialble for UART1 but almost
never used since they also provide the CSI I2C lines.

Signed-off-by: Paul Kocialkowski <contact@paulk.fr>